# swagger.yaml
openapi: 3.0.0
info:
  title: Doctor API
  version: 1.0.0
  description: >
    Doctor API documentation for the Telemedicine system.
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /api/test/encrypt:
    post:
      summary: Encrypt Password
      description: Encrypts a plain text password using Bcrypt and returns the hashed value.
      tags:
        - Test
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  example: mypassword123
                  description: The plain text password to be encrypted.
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  example: mypassword123
                  description: The plain text password to be encrypted.
      responses:
        '200':
          description: Password encrypted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "$2b$14$p.Fv7a5PbQVrwnSgQGFfHe96gpZKB0xRPIY.PN4FvWyfvhHC08gR2"
                    description: The Bcrypt hashed password.
                  status:
                    type: boolean
                    example: true
                    description: Indicates success.
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password is required."
                    description: Error message.
                  status:
                    type: boolean
                    example: false
                    description: Indicates failure.
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "This API is not allowed in production."
                    description: Error message.
                  status:
                    type: boolean
                    example: false
                    description: Indicates failure.
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal server error."
                    description: Error message.
                  status:
                    type: boolean
                    example: false
                    description: Indicates failure.
  /api/test/decrypt:
    post:
      summary: Decrypt Password
      description: Decrypts a plain text password against a Bcrypt hash.
      tags:
        - Test
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
                - hash
              properties:
                password:
                  type: string
                  example: mypassword123
                  description: The plain text password to decrypt.
                hash:
                  type: string
                  example: "$2b$14$p.Fv7a5PbQVrwnSgQGFfHe96gpZKB0xRPIY.PN4FvWyfvhHC08gR2"
                  description: The Bcrypt hashed password to compare against.
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - password
                - hash
              properties:
                password:
                  type: string
                  example: mypassword123
                  description: The plain text password to decrypt.
                hash:
                  type: string
                  example: "$2b$14$p.Fv7a5PbQVrwnSgQGFfHe96gpZKB0xRPIY.PN4FvWyfvhHC08gR2"
                  description: The Bcrypt hashed password to compare against.
      responses:
        '200':
          description: Password decrypted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The password is correct."
                    description: Verification message.
                  status:
                    type: boolean
                    example: true
                    description: Indicates success.
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password and hash are required."
                    description: Error message.
                  status:
                    type: boolean
                    example: false
                    description: Indicates failure.
        '403':
            description: Forbidden
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      example: "This API is not allowed in production."
                      description: Error message.
                    status:
                      type: boolean
                      example: false
                      description: Indicates failure.
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal server error."
                    description: Error message.
                  status:
                    type: boolean
                    example: false
                    description: Indicates failure.
  /api/verifyToken:
    post:
      summary: Verify Token for Website
      description: Verifies the JWT token sent via HttpOnly cookie and returns user info if valid.
      tags:
        - Verify
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  Users_ID:
                    type: integer
                    example: 123
                  Users_Email:
                    type: string
                    example: user@example.com
                  Users_Username:
                    type: string
                    example: username123
                  UsersType_ID:
                    type: integer
                    example: 1
                  Users_Type:
                    type: string
                    example: doctor
                  message:
                    type: string
                    example: Token is valid.
                  status:
                    type: boolean
                    example: true
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid token.
                  status:
                    type: boolean
                    example: false
      security: []
  /api/login:
    post:
      summary: Login Website
      description: Authenticates a doctor or patient user and sets an HttpOnly token cookie
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - Users_Email
                - Users_Password
              properties:
                Users_Email:
                  type: string
                  description: User email or username
                  example: patient@gmail.com
                Users_Password:
                  type: string
                  description: User password
                  example: Cs12345678
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - Users_Email
                - Users_Password
              properties:
                Users_Email:
                  type: string
                  description: User email or username
                  example: patient@gmail.com
                Users_Password:
                  type: string
                  description: User password
                  example: Cs12345678
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: The login was successful.
                  status:
                    type: boolean
                    example: true
          headers:
            Set-Cookie:
              description: HttpOnly cookie containing JWT token
              schema:
                type: string
                example: token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          description: Missing or invalid parameters / Invalid user type
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Please fill in the correct parameters as required.
                  status:
                    type: boolean
                    example: false
        '401':
          description: Email or password incorrect
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Email or password is incorrect.
                  status:
                    type: boolean
                    example: false
        '404':
          description: User type details not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User type details not found.
                  status:
                    type: boolean
                    example: false
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: An error occurred on the server.
                  status:
                    type: boolean
                    example: false
      security: []
  /api/logout:
    post:
      summary: Logout Website
      description: Clears the user's token cookie and logs out the user
      tags:
        - Authentication
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logged out successfully.
                  status:
                    type: boolean
                    example: true
        '401':
          description: Unauthorized / Invalid token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Unauthorized
                  status:
                    type: boolean
                    example: false
      security: []
...
  /api/profile/get:
    get:
      tags:
        - Profile
      summary: Get user profile by token
      description: Retrieve the profile data of a logged-in user (doctor or patient) using the JWT token.
      operationId: getProfileByToken
      parameters: []
      responses:
        '200':
          description: Profile data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Profile data retrieved successfully."
                  Users_Type_Table:
                    type: string
                    example: "patient"
                  Users_Email:
                    type: string
                    example: "user@example.com"
                  Users_Type:
                    type: string
                    enum: [patient, doctor]
                  Patient_ID:
                    type: integer
                    example: 101
                  Doctor_ID:
                    type: integer
                    example: 201
                  Specialty_Name:
                    type: string
                    example: "Cardiology"
        '400':
          description: Missing or invalid user type / ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Missing user type or ID."
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Invalid Token."
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "No profile data found for this user."
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "An error occurred on the server."
    security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        status:
          type: boolean